+++ service/configservice/api_op_PutOrganizationConfigRule.go
@@ -29,24 +29,28 @@ import (
 // administrator, register a delegated administrator by calling Amazon Web Services
 // Organization register-delegated-administrator for
 // config-multiaccountsetup.amazonaws.com. There are two types of rules: Config
-// Custom Rules and Config Managed Rules. You can use PutOrganizationConfigRule to
-// create both Config custom rules and Config managed rules. Custom rules are rules
-// that you can create using either Guard or Lambda functions. Guard (Guard GitHub
-// Repository (https://github.com/aws-cloudformation/cloudformation-guard)) is a
-// policy-as-code language that allows you to write policies that are enforced by
-// Config Custom Policy rules. Lambda uses custom code that you upload to evaluate
-// a custom rule. If you are adding a new Custom Lambda rule, you first need to
-// create an Lambda function in the management account or a delegated administrator
-// that the rule invokes to evaluate your resources. You also need to create an IAM
-// role in the managed account that can be assumed by the Lambda function. When you
-// use PutOrganizationConfigRule to add a Custom Lambda rule to Config, you must
-// specify the Amazon Resource Name (ARN) that Lambda assigns to the function.
-// Managed rules are predefined, customizable rules created by Config. For a list
-// of managed rules, see List of Config Managed Rules
+// Managed Rules and Config Custom Rules. You can use PutOrganizationConfigRule to
+// create both Config Managed Rules and Config Custom Rules. Config Managed Rules
+// are predefined, customizable rules created by Config. For a list of managed
+// rules, see List of Config Managed Rules
 // (https://docs.aws.amazon.com/config/latest/developerguide/managed-rules-by-aws-config.html).
 // If you are adding an Config managed rule, you must specify the rule's identifier
-// for the RuleIdentifier key. Prerequisite: Ensure you call EnableAllFeatures API
-// to enable all features in an organization. Make sure to specify one of either
+// for the RuleIdentifier key. Config Custom Rules are rules that you create from
+// scratch. There are two ways to create Config custom rules: with Lambda functions
+// ( Lambda Developer Guide
+// (https://docs.aws.amazon.com/config/latest/developerguide/gettingstarted-concepts.html#gettingstarted-concepts-function))
+// and with Guard (Guard GitHub Repository
+// (https://github.com/aws-cloudformation/cloudformation-guard)), a policy-as-code
+// language. Config custom rules created with Lambda are called Config Custom
+// Lambda Rules and Config custom rules created with Guard are called Config Custom
+// Policy Rules. If you are adding a new Config Custom Lambda rule, you first need
+// to create an Lambda function in the management account or a delegated
+// administrator that the rule invokes to evaluate your resources. You also need to
+// create an IAM role in the managed account that can be assumed by the Lambda
+// function. When you use PutOrganizationConfigRule to add a Custom Lambda rule to
+// Config, you must specify the Amazon Resource Name (ARN) that Lambda assigns to
+// the function. Prerequisite: Ensure you call EnableAllFeatures API to enable all
+// features in an organization. Make sure to specify one of either
 // OrganizationCustomPolicyRuleMetadata for Custom Policy rules,
 // OrganizationCustomRuleMetadata for Custom Lambda rules, or
 // OrganizationManagedRuleMetadata for managed rules.